feat: 방 이벤트 구독
