+++ src/app/room/[slug]/page.tsx
@@ -1,11 +1,14 @@
 "use client";
 
-import { useEffect, useRef, useState } from "react";
+import { useCallback, useEffect, useRef, useState } from "react";
+import type { StompSubscription } from "@stomp/stompjs";
 import { useParams } from "next/navigation";
 import {
   joinRoom,
   type JoinRoomResult,
 } from "@/src/entities/room/api/joinRoom";
+import { subscribeRoomEvents } from "@/src/entities/room/api/websocket/subscribeRoomEvents";
+import type { WsEvent } from "@/src/entities/room/model/types";
 import { ApiError } from "@/src/shared/api/api-error";
 import RoomPasswordInput from "@/src/features/room/join/ui/roomPasswordInput";
 
@@ -18,11 +21,59 @@ export default function RoomPage() {
     slug: string;
     promise: Promise<JoinRoomResult>;
   } | null>(null);
+  const roomSubscriptionRef = useRef<{
+    slug: string;
+    subscription: StompSubscription;
+  } | null>(null);
 
   const [status, setStatus] = useState<JoinStatus>("joining");
   const [message, setMessage] = useState("joining...");
   const [code, setErrorCode] = useState("joining...");
   const [isSubmittingPassword, setIsSubmittingPassword] = useState(false);
+  const [lastRoomEventType, setLastRoomEventType] = useState("없음");
+  const [lastRoomEventTime, setLastRoomEventTime] = useState("");
+
+  const cleanupRoomSubscription = useCallback(() => {
+    if (!roomSubscriptionRef.current) {
+      return;
+    }
+
+    try {
+      roomSubscriptionRef.current.subscription.unsubscribe();
+    } catch {
+      // The socket may already be closing while the page is leaving.
+    }
+
+    roomSubscriptionRef.current = null;
+  }, []);
+
+  const ensureRoomSubscription = useCallback(
+    (roomSlug: string) => {
+    if (roomSubscriptionRef.current?.slug === roomSlug) {
+      return;
+    }
+
+    cleanupRoomSubscription();
+
+    roomSubscriptionRef.current = {
+      slug: roomSlug,
+      subscription: subscribeRoomEvents(roomSlug, ({ body }) => {
+        if (!body) return;
+
+        let event: WsEvent;
+        try {
+          event = JSON.parse(body) as WsEvent;
+        } catch {
+          return;
+        }
+
+        setLastRoomEventType(event.type);
+        setLastRoomEventTime(new Date(event.timestamp).toLocaleTimeString());
+      }),
+    };
+    },
+    [cleanupRoomSubscription],
+  );
 
   async function handlePasswordSubmit(password: string) {
     if (!slug) return;
@@ -32,6 +83,7 @@ export default function RoomPage() {
 
     try {
       const result = await joinRoom(slug, { password });
+      ensureRoomSubscription(slug);
       setStatus("joined");
       setMessage(
         `joined at ${new Date(result.timestamp).toLocaleTimeString()}`,
@@ -73,6 +125,7 @@ export default function RoomPage() {
         const result = await currentJoinRequest.promise;
         if (!isActive) return;
 
+        ensureRoomSubscription(slug);
         setStatus("joined");
         setMessage(
           `joined at ${new Date(result.timestamp).toLocaleTimeString()}`,
@@ -95,8 +148,9 @@ export default function RoomPage() {
 
     return () => {
       isActive = false;
+      cleanupRoomSubscription();
     };
-  }, [slug]);
+  }, [slug, cleanupRoomSubscription, ensureRoomSubscription]);
 
   if (status === "needs-password") {
     return (
@@ -114,6 +168,8 @@ export default function RoomPage() {
       <div>join: {status}</div>
       <div>message: {message}</div>
       <div>Error Code: {code}</div>
+      <div>last room event: {lastRoomEventType}</div>
+      <div>last room event time: {lastRoomEventTime || "-"}</div>
     </div>
   );
 }
