+++ src/entities/room/api/joinRoom.types.ts
@@ -1,0 +1,9 @@
+export type JoinRoomPayload = {
+  password?: string | null;
+};
+
+export type JoinRoomResult = {
+  roomSlug: string;
+  timestamp: number;
+  data: unknown;
+};
